updating c++ code to ansi c. Initial commit
