+++ imageutils/laxwrappers.pyx
@@ -1,6 +1,3 @@
-# distutils: language = c++
-# distutils: sources = laxutils.cpp
-# cython: profile=True
 import numpy as np
 cimport numpy as np
 cimport cython
